requirements.txt
pandas
numpy
requests
matplotlib
scikit-learn
yfinance
langchain-openai